Separated examples from library code.

+++ test/CPFDSpec.hs
@@ -1,17 +1,17 @@
 module CPFDSpec where
 
 import Control.CPFD
-import Test.Hspec
-import Test.Hspec.QuickCheck (prop)
-import Test.QuickCheck hiding ((.&.))
 import Data.Maybe (maybeToList)
+import Test.Hspec
+-- import Test.Hspec.QuickCheck (prop)
+-- import Test.QuickCheck hiding ((.&.))
 
 
 spec :: Spec
 spec = do
 
   describe "List#cmap" $ do
-    let cMaybe = List [Just 123, Nothing]
+    let cMaybe = List [Just (123::Int), Nothing]
     let cList  = List [[123], []]
     it "id" $
       cmap id cMaybe `shouldBe` cMaybe
